Revise, update, and expand several Rationale entries.
Note that this removes the "The history of ```nop``` instructions"
sentence, because the long history of ```nop``` instructions is mostly
about uses that don't apply to wasm: alignment padding, hazard avoidance,
vliw packing, branch delay slot stuffing, timing, and modifying code at
runtime. The one use that does apply to wasm is already described in the
preceding paragraph.